requirements.txt
requests
python-dotenv
cloudinary